Fix: Do not use OutputPage::$mBodytext in SkinTemplateOutputPageBeforeExec
The bodytext data of the Template is not set by reference to OutputPage::$mBodytext
anymore but is now added by value. CHanges to mBodytext will therefore be not re-
flected in the template anymore. Use the bodytext data of the Template instead.

ALso: Remove the use MediaWikiUI variable, as it doesn't need to be set anymore
(InputBox uses OOUI by default).