Re-enable mobile ads after the content
This was probably broken by because of the use of Skin::getOutput()->addHTML,
which will add the content to the OutputPage but the content of OutputPage was
already added to the Template, but by value not by reference. That's why it
is not updated anymore. Use the onSkinTemplateOutputPageBeforeExec hook handler
to add mobile ads, too.